requirements.txt
python-twitter-v2==0.8.1
python-dateutil==2.8.2
PyYAML==6.0
schema==0.7.5
rfc3339==6.2
